Update SchoolClass.java
In the loop condition, pupils.size() is re-evaluated in each iteration, and because the code removed pupils from the list inside the loop, the size of the list decreases with each iteration. This is not clean and because of this an exception occurred when trying to run the function with an even amount of pupils. 

Furthermore the if condition to check if the size of the array is 1 needed also to be inside the loop for this solution.